requirements.txt
tensorflow>=2.4.0
numpy>=1.19.2
matplotlib>=3.3.2
scikit-learn>=0.24.0
Pillow>=8.0.0
streamlit>=1.10.0
opencv-python>=4.5.1